refactor(logging): add log package under core
alter the location of some functions
