+++ src/main/core/log/safe-stringify.ts
