+++ src/main/core/scope.ts
@@ -4,7 +4,7 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-import { Loggable } from './loggable.js'
+import { Loggable } from './log/loggable.js'
 import { type ScopeMetric } from './scope-metric.js'
 
 /**
